refactor: unified linalg and elementwise headers into tensr_math

+++ src/tensr/math/elementwise/tensr_sqrt.c
@@ -6,7 +6,7 @@
 /*   By: sgadinga <sgadinga@student.42abudhabi.ae>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/06 01:57:22 by sgadinga          #+#    #+#             */
-/*   Updated: 2026/03/06 02:42:17 by sgadinga         ###   ########.fr       */
+/*   Updated: 2026/04/19 19:47:40 by sgadinga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ t_tensr	*tensr_sqrt(const t_tensr *t, t_tensr *out)
 		src = tensr_get(t, it.indices);
 		dst = tensr_get(out, it.indices);
 		if (out->dtype == DT_F32)
-			*(float *)dst = sqrt(*(float *)src);
+			*(float *)dst = sqrtf(*(float *)src);
 		else if (out->dtype == DT_F64)
 			*(double *)dst = sqrt(*(double *)src);
 	}
